# Keep connections cached, and reduce round-trips