--- a/src/include/storage/postgres_schema_set.hpp
+++ b/src/include/storage/postgres_schema_set.hpp
@@ -21,6 +21,8 @@ public:
 public:
 	optional_ptr<CatalogEntry> CreateSchema(ClientContext &context, CreateSchemaInfo &info);
 
+	static string GetInitializeQuery();
+
 protected:
 	void LoadEntries(ClientContext &context) override;
 };