--- a/src/include/storage/postgres_transaction.hpp
+++ b/src/include/storage/postgres_transaction.hpp
@@ -32,6 +32,7 @@ public:
 		return connection.GetDSN();
 	}
 	unique_ptr<PostgresResult> Query(const string &query);
+	vector<unique_ptr<PostgresResult>> ExecuteQueries(const string &queries);
 	static PostgresTransaction &Get(ClientContext &context, Catalog &catalog);
 
 private: