--- a/src/include/storage/postgres_transaction.hpp
+++ b/src/include/storage/postgres_transaction.hpp
@@ -30,8 +30,6 @@ public:
 
 	PostgresConnection &GetConnection();
 	string GetDSN();
-	//! Retrieves the connection **without** starting a transaction if none is active
-	PostgresConnection &GetConnectionRaw();
 	unique_ptr<PostgresResult> Query(const string &query);
 	vector<unique_ptr<PostgresResult>> ExecuteQueries(const string &queries);
 	static PostgresTransaction &Get(ClientContext &context, Catalog &catalog);
@@ -40,6 +38,10 @@ private:
 	PostgresPoolConnection connection;
 	PostgresTransactionState transaction_state;
 	AccessMode access_mode;
+
+private:
+	//! Retrieves the connection **without** starting a transaction if none is active
+	PostgresConnection &GetConnectionRaw();
 };
 
 } // namespace duckdb