--- a/src/include/storage/postgres_type_set.hpp
+++ b/src/include/storage/postgres_type_set.hpp
@@ -19,12 +19,12 @@ struct PGTypeInfo;
 
 class PostgresTypeSet : public PostgresCatalogSet {
 public:
-	explicit PostgresTypeSet(PostgresSchemaEntry &schema);
+	explicit PostgresTypeSet(PostgresSchemaEntry &schema,
+							 unique_ptr<PostgresResultSlice> enum_result = nullptr,
+							 unique_ptr<PostgresResultSlice> composite_type_result = nullptr);
 
 public:
 	optional_ptr<CatalogEntry> CreateType(ClientContext &context, CreateTypeInfo &info);
-
-	void Initialize(PostgresTransaction &transaction, PostgresResultSlice &enums, PostgresResultSlice &composite_types);
 
 	static string GetInitializeEnumsQuery();
 	static string GetInitializeCompositesQuery();
@@ -38,9 +38,10 @@ protected:
 	void InitializeEnums(PostgresResultSlice &enums);
 	void InitializeCompositeTypes(PostgresTransaction &transaction, PostgresResultSlice &composite_types);
 
-
 protected:
 	PostgresSchemaEntry &schema;
+	unique_ptr<PostgresResultSlice> enum_result;
+	unique_ptr<PostgresResultSlice> composite_type_result;
 };
 
 } // namespace duckdb