--- a/src/postgres_query.cpp
+++ b/src/postgres_query.cpp
@@ -64,7 +64,7 @@ static unique_ptr<FunctionData> PGQueryBind(ClientContext &context, TableFunctio
 
 	// set up the bind data
 	result->dsn = con.GetDSN();
-	result->connection = PostgresConnection(con.GetConnection());
+	result->SetConnection(con.GetConnection());
 	result->types = return_types;
 	result->names = names;
 	result->read_only = false;