--- a/src/postgres_storage.cpp
+++ b/src/postgres_storage.cpp
@@ -9,9 +9,7 @@ namespace duckdb {
 
 static unique_ptr<Catalog> PostgresAttach(StorageExtensionInfo *storage_info, AttachedDatabase &db, const string &name,
                                         AttachInfo &info, AccessMode access_mode) {
-	auto connection = PostgresConnection::Open(info.path);
-	auto version = connection.GetPostgresVersion();
-	return make_uniq<PostgresCatalog>(version, db, info.path, access_mode);
+	return make_uniq<PostgresCatalog>(db, info.path, access_mode);
 }
 
 static unique_ptr<TransactionManager> PostgresCreateTransactionManager(StorageExtensionInfo *storage_info,