--- a/src/storage/postgres_catalog.cpp
+++ b/src/storage/postgres_catalog.cpp
@@ -9,13 +9,16 @@
 
 namespace duckdb {
 
-PostgresCatalog::PostgresCatalog(PostgresVersion version, AttachedDatabase &db_p, const string &path, AccessMode access_mode)
-    : Catalog(db_p), path(path), access_mode(access_mode), version(version), schemas(*this) {
+PostgresCatalog::PostgresCatalog(AttachedDatabase &db_p, const string &path, AccessMode access_mode)
+    : Catalog(db_p), path(path), access_mode(access_mode), schemas(*this), connection_pool(*this) {
 	Value connection_limit;
 	auto &db_instance = db_p.GetDatabase();
 	if (db_instance.TryGetCurrentSetting("pg_connection_limit", connection_limit)) {
 		connection_pool.SetMaximumConnections(UBigIntValue::Get(connection_limit));
 	}
+
+	auto connection = connection_pool.GetConnection();
+	this->version = connection.GetConnection().GetPostgresVersion();
 }
 
 PostgresCatalog::~PostgresCatalog() = default;