--- a/vcpkg.json
+++ b/vcpkg.json
@@ -1,5 +1,8 @@
 {
   "dependencies": [
     "openssl"
+  ],
+  "overrides": [
+    {"name": "openssl", "version": "3.0.8"}
   ]
 }